feat: add user register method

+++ src/utils/fileUpload.js
@@ -5,25 +5,27 @@ import fs from "fs";
 
 cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-    api_secret: CLOUDINARY_API_SECRET,
-    api_key: CLOUDINARY_API_KEY
+    api_secret: process.env.CLOUDINARY_API_SECRET,
+    api_key: process.env.CLOUDINARY_API_KEY
 })
 
-const uploadOnCloudinary = async (localFilePath) => {
+export const uploadOnCloudinary = async (localFilePath) => {
     try{
         if (!localFilePath) return null
 
         // upload file on cloudinary
-        cloudinary.uploader.upload(localFilePath, {
+        const response = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto"
         })
-
         // file has been uploaded to cloudinary
         console.log("File has been uploaded to cloudinary: ", response.url)
         return response;
     }
     catch (error) {
         // remove the locally saved temporary file as upload got failed
-        fs.unlinkSync(localFilePath)
+        console.error("Error while uploading file: ", error.message)
+    }
+    finally{
+        if(localFilePath) fs.unlinkSync(localFilePath)
     }
 }
